Update and rename LICENSE to CODE-LICENSE.txt
The source code and the artwork are under different licenses.

The source code is open source and distributed under the "Unlicense" license, which is effectively public domain.

The artwork, however, is copyright Quixotic Games and is used by permission.  It is all taken from the print-and-play version of the game.