Use more muted player colors
Using pure versions of red, blue, green, and yellow player pieces
is a bit jarring, so use shades of these colors that match the
player boards.

Except for yellow.  Unfortunately, a matching shade of yellow is
too similar to the colors of the game board, making yellow cubes
and tokens too hard to see.

Fixes #1