give City Well a purple, not red, border
In the board game, the City Well token has a red border, but that's
confusing because there is a red player.  Since the TTS implementation
gives each tile the edge color of the player, having a red border
and purple edge is even worse.  So replace the red border with a
purple one.

Fixed #3