Set the turn order when the game starts
When the Start Game button is pressed, the turn order in the game
should be set.  In a real multi-player game, there's no real need
to maintain turn order.  In hot-seat mode, turn order does matter.

Also, after the game is started, the Start Game button is deleted
so that you don't accidentally try to start the game again.